+++ src/main/java/domain/player/Name.java
@@ -1,4 +1,4 @@
-package domain.participant;
+package domain.player;
 
 public class Name {
